requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.0.2
cssselect==1.0.1
Django==1.11.4
enum34==1.1.6
hyperlink==17.3.0
idna==2.5
image==1.5.11
incremental==17.5.0
ipaddress==1.0.18
lxml==3.8.0
MySQL-python==1.2.5
olefile==0.44
parsel==1.2.0
Pillow==4.2.1
pkg-resources==0.0.0
pyasn1==0.3.1
pyasn1-modules==0.0.10
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.2.0
pytz==2017.2
queuelib==1.4.2
requests==2.18.3
rsa==3.4.2
Scrapy==1.4.0
selenium==3.4.3
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
urllib3==1.22
w3lib==1.17.0
zope.interface==4.4.2
